requirements.txt
# Core dependencies
pyyaml>=6.0.1
pydantic>=2.5.0
python-dotenv>=1.0.0

# Messaging
kafka-python>=2.0.2
pika>=1.3.2

# Databases
psycopg2-binary>=2.9.9
elasticsearch>=8.11.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.2

# SFTP
paramiko>=3.4.0

# Caching
redis>=5.0.1

# Testing & Reporting
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-html>=4.1.1
jinja2>=3.1.2

# Monitoring & Metrics
prometheus-client>=0.19.0

# CLI
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0

# Async
asyncio>=3.4.3
aiohttp>=3.9.1

# Logging
structlog>=23.2.0
python-json-logger>=2.0.7

# Security
cryptography>=41.0.7
pyjwt>=2.8.0

# Utils
tenacity>=8.2.3
python-dateutil>=2.8.2
